[router] fix passing style array to <Link /> when using asChild (#31373)
# Why

Fix: #31352

RadixUI has special logic to handle the merging of `style` and
`className` props. On the web styles are not allowed so Radix does not
handle this scenario/ This could be fixed upstream (PR open), but it may
not as RN is not their target platform.


# How

This shim calls `StyleSheet.flatten` on the styles before we render the
<Slot />

We will need to do the same for the Headless components which use `<Slot
/>` as well.

# Test Plan

<!--
Please describe how you tested this change and how a reviewer could
reproduce your test, especially if this PR does not include automated
tests! If possible, please also provide terminal output and/or
screenshots demonstrating your test/reproduction.
-->

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
This is required for changes to Expo modules.
-->

- [ ] Documentation is up to date to reflect these changes (eg:
https://docs.expo.dev and README.md).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>