[router]: add new headless Tabs component (#30767)
# Why

 Create new un-styled navigator components for Expo Router.

# How

Create new package export `expo-router/headless` 

## Components

```
<Tabs>
      <TabSlot /> // Renders the content of the tab
      <TabList> // Container view for the tab bar.
        <TabTrigger href="/ > // Tab pressable
          <Text>Home</Text>
        </TabTrigger>
        <TabTrigger href="/article">
            <Text>article</Text>
        </TabTrigger>
      </TabList>
    </Tabs>
```

`TabTrigger` supports the `asChild` props (similar to `<Link />` to
allow for custom styled triggers.

<hr />

### `<Tabs>`

The container tabs. Internally this calls `useNavigationBuilder` and
converts the children to `<Screen />`

```
type TabsProps = ViewProps & {
  options?: UseTabsOptions
};

// UseTabsOptions are the options passed to useNavigationBuilder, except initialRouteName and children
```

### `<TabSlot>`

```
export type TabSlotProps = ViewProps & {
  asChild?: boolean;
  options?: UseTabSlotOptions;
};
```

A `<Slot />` component that defaults to a `<View />`. A custom slot can
be made with the `useTabSlot()` hook.

### `<TabList>`

A `<Slot />` component that defaults to a `<View />`. Only used as a
wrapper for `<TabTriggers />`

### `<TabTrigger>`

A `<Slot />` component that defaults to a `<Link />`.

## Hooks

`useTabsWithChildren({ children })`

Can be used to create custom `<Tabs />` components. Will parse the
`children` for triggers

`useTabsWithTriggers({ triggers })`

Can be used to create custom `<Tabs />` components. Triggers need to be
explicitly passed. Useful when you need to customise how the `children`
are rendered

`useTabsSlot(options?: UseTabsSlotOptions)`

```
export type UseTabsSlotOptions = {
  detachInactiveScreens?: boolean;
  renderFn?: ((descriptor: TabsDescriptor, { isFocused, loaded }) => React.JSX.Element) | undefined
};
```

A hook to render the current slot. Can pass a custom render function to
customise the `<Screen />` (this is a `react-native-screens` screen)

# Test Plan

This is just the initial implementation with a unit test. Full tests
coming in a later PR

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
This is required for changes to Expo modules.
-->

- [ ] Documentation is up to date to reflect these changes (eg:
https://docs.expo.dev and README.md).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>
Co-authored-by: Evan Bacon <bacon@expo.io>