feat(cli): disable cache reset in CI (#31663)
# Why

The React Native build scripts always enable the cache reset but we
shouldn't need this in CI environments. By disabling it, we can eagerly
bundle code before the build and reuse the cached artifacts in
subsequent builds.


<!--
Please describe the motivation for this PR, and link to relevant GitHub
issues, forums posts, or feature requests.
-->

# How

When CI=1, preserve the export:embed cache. This is primarily designed
for local builds where a cache can be preserved between builds and may
not update (e.g. when changing the babel config).

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>
Co-authored-by: Cedric van Putten <me@bycedric.com>