refactor(cli): clean up the old `UserSettings` file (#31280)
# Why

This cleans up the `UserSettings` file, which was copied over from the legacy `expo-cli`. It exports individual methods and does not mutate the `getSettings` function to add all methods.

I did this in advance of the telemetry refactor, which will be the next stacked PR.

# How

- Refactored `UserSettings` without `Object.assign` and `export {}`

# Test Plan

See tests

# Checklist

<!--
Please check the appropriate items below if they apply to your diff. This is required for changes to Expo modules.
-->

- [x] Documentation is up to date to reflect these changes (eg: https://docs.expo.dev and README.md).
- [ ] Conforms with the [Documentation Writing Style Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build (eg: updated a module plugin).