refactor(cli): add `instant` telemetry strategy using granular `fetch` (#31281)
# Why

This refactors the telemetry utility and splits it up into 3 different
strategies:

1. `detached` → Store all telemetry events in memory, and submit them in
a separate process once the CLI closes using the rudder client
2. `debug` → Use the rudder client directly to send telemetry events,
this library is slow and uses batching logic that's created for servers,
which may not work as effectively in CLIs. This is enabled through
`EXPO_NO_TELEMETRY_DETACH`
3. `instant` → (new) Use `fetch` (not the rudder client) to instantly
submit telemetry events. All events are retried 3 times, with a short
timeout on connection. This is great for long-running commands, like
`start`, `export*` and `run*`.

In the future, we might be able to drop the rudder client in favor of a
batched `FetchClient` (`instant` strategy). The most important part of
the `FetchClient` is that it has fine-grained control over every
individually sent events.

# How

- Added distinct clients for each telemetry strategy
- Added ability to switch strategies, without losing telemetry events
(except for rudder, that's not possible as we have no control over this)

# Test Plan

See tests

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
This is required for changes to Expo modules.
-->

- [x] Documentation is up to date to reflect these changes (eg:
https://docs.expo.dev and README.md).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).