Android 14: Register Screen Capture callback earlier
On Android 14 we're using a new ScreenCapture detection API which requries an extra step in our module to register the callback. Due to a bug in bridgeless mode the activity wasn't ready when the module was created leading to a crash if we tried to register the callback here.

A fix was made in #28244 where the registration code was moved into the API methods for preventing/allowing screen capture. This was fine in the example in our tests since we use these methods in our test code - however in #31678 the example (from documentation) used did not automatically call the required API methods and so the callback was not registered and screen capture notifications was not sent on Android 14.

This PR fixes this by trying to register in the `onCreate` method (testing for null) in addition to also registering in `onActivityEntersForeground` as a fallback.

Registering will no longer be done in the `allowScreenCapture` and `preventScreenCapture` API methods. Registration also has an extra null check.

Related PRs: #28244 which was waiting on #28200 - both merged.