feat(template): add blur tab bar on iOS (#31739)
# Why

- It's common to make the tab bar on iOS be blurry so you can see
content scroll under it. But doing so requires extra changes to the root
scroll view to move the padding up. This of course, doesn't work well on
Android or web, so it's iOS only. But it'll encourage more native
looking UI in apps.
- Also adds an otherwise undocumented note about using
`systemChromeMaterial` which adapts to the native appearance
automatically.