refactor(cli): avoid retrying `development-sessions` requests (#31772)
# Why

If the server is unavailable, or something else is going wrong in the
development session calls, do not retry. It's on a best effort basis,
and if it fails, there isn't a lot we can do.

# How

- Dropped `onError` callback in `DevelopmentSession`

# Test Plan

- In `packages/@expo/cli/src/api/updateDevelopmentSession.ts`
- Create a module variable, like `let calledCount = 0;`
- Modify both `updateDevelopmentSessionAsync` and
`closeDevelopmentSessionAsync`
- Add to the top of the function: `if (calledCount <= 50) throw new
Error('Test')`
- `DEBUG="expo:start:server:developmentSession" expod start

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
This is required for changes to Expo modules.
-->

- [ ] Documentation is up to date to reflect these changes (eg:
https://docs.expo.dev and README.md).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).