feat(cli): link server deployment to client (#31707)
# Why

- React Native apps always use a server in development meaning they can
make requests for static assets and interact with development servers,
but this breaks automatically in production. This PR aims to reconcile
the differences in release builds.
- When `output: "server"` is enabled (during eager bundles, in release
mode, with network connection), a standalone version of the server will
be exported and uploaded to a remote origin.
- The draft URL will be added to the `extra.router.generatedOrigin`
field, which will be used in the `@expo/metro-runtime` fetch polyfill to
make relative requests for assets in the `/public` directory and API
routes.
- If a manual `origin` is used, then it will take priority over the
generated value, allowing for full customization using environment
variables and `app.config.js`.
- If a preset "native:deploy" (`native:deploy:ios`,
`native:deploy:android`) `script` is found in the project's
`package.json`, then this will be used for deploying the server.
Otherwise a default command will be used instead.
- Conformance: the replacement script must support the `--export-dir
<absolute-path>` argument, and log `{ url: "http://output-url.com" }` to
the stderr.
- This feature can be disabled with `EXPO_OFFLINE=1`, or the new
`EXPO_NO_DEPLOY=1`. It will only run via `npx expo run` commands, and
EAS Build, and only for production builds. Building from Xcode and
Android Studio will work the same as before as they are not capable of
spawning scripts that have full access to the user's SHELL.
- The restrictions on `origin` have been loosened to account for this
new default flow. Meaning you can always perform relative requests in
development, as the polyfill assumes that you'll

## TODO

- [ ] Mitigate `output: "static"` differences.
- [ ] Docs

# Test Plan

- [x] Manual verification.
- TBD

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>