[prebuild-config] add support for iOS 18 dark/tinted icons (#30247)
iOS 18 introduced support for dark/tinted icons in the launcher/home
screen. This commit introduces an opt-in, backwards-compatible pattern for
users to specify a dark/tinted variant of their icon in case they happen
to build their apps for iOS 18. Building on current EAS infrastructure
(Xcode 15) is not affected, it still works fine.
The implementation has been tested on both Xcode 15 and 16

Co-authored-by: Zac Adams <77133479+zacherya@users.noreply.github.com>
Co-authored-by: Brent Vatne <brentvatne@gmail.com>