Adds more context to build and update docs on home page (#31674)
# Why

@brentvatne mentioned that the build project and over-the-air update doc
could be improved with more content. As a result of that conversation,
this PR:

- Makes the "configure updates" section of the build-project doc
collapsible and clearly labels it as optional
- Adds documentation on how to build locally without EAS CLI.
- Adds a "Learn more" section to the bottom of both docs that links to
their sections in /guides. I didn't use the "Next Step" header since
there are buttons that invite users to the next doc as the next step,
and I'd like to have one primary next step on each doc. To make the
"Learn more" a bit more juicy, I added some suggestions on things you
can learn if you go there.

# Test Plan

Read the docs and let me know what you think. I am open to feedback!

I am assuming that the majority of users reading these are closer to the
"trying to get something done" part of the spectrum rather than the
"understanding how everything works" part of the spectrum. They are
newer to Expo's tools and looking for which commands they'll need to
run. --With that, we want them to know there's a lot you can do with
Expo, which we have many detailed guides on.

-
http://docs.expo.dev-pr-31674.s3-website-us-east-1.amazonaws.com/deploy/build-project/
-
http://docs.expo.dev-pr-31674.s3-website-us-east-1.amazonaws.com/deploy/send-over-the-air-updates/

---------

Co-authored-by: Aman Mittal <amandeepmittal@live.com>