[video] Add `timeUpdate` event (#31327)
# Why

Currently to observe the progress of the player a the `currentTime`
property has to be read in JS interval. This PR adds a
`progressUpdateInterval` property and `progressUpdate` event, which can
be used to receive player progress update events.

# How

On Android used `Handler.postDelayed` to create an interval clock, on
iOS used `addPeriodicTimeObserver` on web used a regular JS Interval.

# Test Plan

Tested in BareExpo on Android, iOS and web.