[expo] add a default AppEntryNotFound component (#31813)
# Why

close ENG-12827

# How

in case people has error in `import App from './App` and not getting a
chance to call `registerRootComponent`. then the `Invariant Violation:
"main" has not been registered.` is quite confused and people may not
focus on the real error. this pr tries to prevent the invariant error
and add a default error instead.

---------

Co-authored-by: Tomasz Sapeta <tsapeta@users.noreply.github.com>
Co-authored-by: Brent Vatne <brentvatne@gmail.com>
Co-authored-by: James Ide <ide@users.noreply.github.com>