[live-photo][ios] Initial version of expo-live-photo (#31193)
# Why

ENG-12860
Created an initial version of expo-live-photo

# How

For now only used the iOS API, we can also add web support. For SDK53 we
can research if it's possible to also support Android versions of live
photos (if they are standardized)

The playback can be controlled via functions, it is also possible to
enable/disable the default gesture recogniser. There are also loading
and playback events.

# Test Plan

Tested in BareExpo on iOS



https://github.com/user-attachments/assets/90ee0f4a-788f-4b83-a13f-de66765c675c