remove default babel config (#31928)
# Why

We haven't needed a babel.config.js file with defaults for a while now.
Because Metro doesn't consider the contents of the babel.config.js in
the transform cache, it can lead to unexpected errors if the user
customizes the file without a reasonable understanding in how the
bundler works.

The primary uses of the `babel.config.js` in the past have been to add
TS paths support, environment variables, and reanimated. All of these
features are built-in to Expo CLI by default now.

Users can still expose the babel config by running `npx expo customize
babel.config.js`.