feat(metro-config)!: Emit assets as strings instead of numeric lookups on web (#31955)
# Why

The non-standard Metro asset format can be confusing to work with on
web. Metro turns assets into `number` which maps to an ID that can be
used to lookup a manifest of asset info at runtime. This manifest is
used to dynamically determine which resolution of asset to use along
with the local file location when the bundle is embedded.

None of this is really needed on web, in dom components, or with server
rendering.

By removing support, this also means we remove support for
multi-platform asset extensions on web. Devs should instead use
`<picture>` element or related browser mechanisms to replicate this
functionality in a more meaningful way.

We'll still use the previous system with native RSC for now but we'll
want to revisit in the future to further optimize how server rendering
works on native.

If this becomes a bigger problem, we can investigate adding the numeric
ID back when multiple scales of an image are detected.

- ~~Another major breaking change (that may be blocking) is that local
images are not measured anymore, meaning image layout will behave
differently on web. Investigating a way around this.~~
- We'll now output images in object format `{ uri: string, width:
number, height: number }` which is supported in React Native web and
resembles what other web frameworks do.

# Test Plan

- Added more asset transformer tests.
- Existing E2E tests should continue to work.

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>