[notifications][android] fix notification with ChannelAwareTrigger not being presented (#31999)
# Why

When presenting a notification that specifies `channelId` in the trigger (as seen in the example code added), calling `scheduleNotificationAsync` would reject.

# How

Handled the case of `request.trigger is ChannelAwareTrigger` specifically.

It's not a nice fix, and I believe `channel` shouldn't be part of a trigger at all (`ChannelAwareTrigger` wouldn't exists) but this is the least invasive fix

# Test Plan

tested with the added example code in Expo Go

# Checklist

<!--
Please check the appropriate items below if they apply to your diff. This is required for changes to Expo modules.
-->

- [ ] Documentation is up to date to reflect these changes (eg: https://docs.expo.dev and README.md).
- [ ] Conforms with the [Documentation Writing Style Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build (eg: updated a module plugin).