[video][android] Audit Picture in Picture on Android (#32043)
# Why

There are some issues with PiP mode android: 

- Fullscreen with startsPictureInPictureAutomatically mismatch -
https://github.com/expo/expo/issues/31916
- Picture in Picture would pause on auto-enter
- `Activity` would keep auto-entering Picture in Picture after leaving
the screen with a video
- Sometimes when entering fullscreen the `MainActivity` would enter
PictureInPicture. This is hard to reproduce, setting `isFullscreen` just
before starting the fullscreen activity should solve the issue though.


# How

- Added PiP support to the `FullscreenPlayerActivity.kt`, managed to
create some common helper functions to simplify the code.
- Picture in Picture would pause on auto-enter - There was no easy
solution for this - there is currently no way to detect when the PiP
transition starts (except for Android SDK 35), so we have to unpause the
player is Activities `onPause` paused it. Thankfully the 1 frame pause
is not noticeable when testing.
- `Activity` would keep auto-entering Picture in Picture after leaving
the screen with a video - when the view detaches from a window
`enterPictureInPictureAutomatically` property of the Activity is
disabled

Also reworked how the `rectHint` for the transition is calculated. The
new way is much simpler and universal and just as reliable (as reliable
as the PiP API allows us to be (read the comment in
`PictureInPictureUtils.kt`))

The code is a bit of a flag-spaghetti at times, this is because the PiP
API on Android is honestly awful, the end effect for the end users seems
pretty smooth though.

# Test Plan

Tested in BareExpo on a SDK 34 emulator and device