chore(expo-dev-menu,expo-dev-launcher): Remove unused `semver` dependencies (#32063)
# Why

I've discovered this by chance. Not sure how best to test this, but it
seems the dependency on `semver` in `expo-dev-launcher` and
`expo-dev-menu` isn't used anymore.

It seems that it was introduced in this change:
https://github.com/expo/expo/commit/ca2c0598d2fce898402543e7fb5d6a7b9a399214
/ #16495

And later this change removed its usage:

https://github.com/expo/expo/commit/7f9ebcf161377ebc8ef400f172c6a0a5e7ceb03a
/ #24504

But the dependency was never removed.

On a side-note, there's also `resolve-from`. I haven't yet looked into
where that's been added and then removed but it also seems to be unused.
I've omitted it from this PR though for now.

# How

- Remove `semver` dependency from the two packages

# Test Plan

Since this only affects the plugin:
- Copy over / Pack & Install `expo-dev-launcher` and `expo-dev-menu`
into a new Expo project
- Add resolutions accordingly and ensure `plugin/build` is up-to-date
with this PR
- Run `expo prebuild`

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
This is required for changes to Expo modules.
-->

- [ ] Documentation is up to date to reflect these changes (eg:
https://docs.expo.dev and README.md).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>