fix(cli): only test if web is supported through `react-dom` (#32148)
# Why

This changes the `WebSupportProjectPrerequisite` to only test on `react-dom`, not `react-native-web`. Testing on `react-native-web` is irrelevant as an Expo project should be able to run without `react-native-web`, when not importing any of `react-native` (like the fixture from #32145).

# How

- Removed `react-native-web` as prerequisite for web support.

# Test Plan

See E2E test from #32145 

# Checklist

<!--
Please check the appropriate items below if they apply to your diff. This is required for changes to Expo modules.
-->

- [x] Documentation is up to date to reflect these changes (eg: https://docs.expo.dev and README.md).
- [ ] Conforms with the [Documentation Writing Style Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build (eg: updated a module plugin).