fix(config): enable web when `react-dom` is installed (#32149)
# Why

This changes the default enabled platforms to include `web`, if `react-dom` is installed. Previously, we checked for `react-native-web`, but that's irrelevant when not importing any `react-native` APIs on web.

# How

- Fixed `getSupportedPlatforms` to test on `react-dom`

# Test Plan

See E2E test from #32145

# Checklist

<!--
Please check the appropriate items below if they apply to your diff. This is required for changes to Expo modules.
-->

- [x] Documentation is up to date to reflect these changes (eg: https://docs.expo.dev and README.md).
- [ ] Conforms with the [Documentation Writing Style Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build (eg: updated a module plugin).