Fix <Slot /> inside headless <Tabs /> causing invalid navigation state  (#32114)
# Why

Ensure that the headless `<Tabs />` use the same navigation context as
the other navigators.

This fixes an issue where a user's mixing `<TabSlot />` with `<Slot />`,
where it would create two navigators. Now it will display the `<Tab />`,
but none of the tab information will be available to the screen (e.g
`isFocused`)

In the future we may want to simplify the API so there is only `<Slot
/>`, and `<Tabs />` has a way to override how `<Slot />` is rendered.

# How

`<Slot />` uses a navigation context to determine if it is a navigator
or should just render a descriptor. `<Tabs />` now uses the same
context.

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>