[router]: add sitemap option to disable sitemap generation (#31701)
# Why

Not every website requires a sitemap (or they may generate their own
custom one). This option allows users to disable this setting.

# How

Add a new config plugin setting `sitemap`. When set to `false` the
sitemap will, not be generated.

I considered making this an option to disable the other generated files,
but they don't make sense to disable.

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>
Co-authored-by: Aman Mittal <amandeepmittal@live.com>