[Apple TV] Support expo-linking package (#32255)
# Why

Testing with canary builds from main shows that Apple TV apps using Expo Router are broken, because expo-linking is now required within the Expo Router code, and expo-linking was not being included in tvOS builds (even though our docs say it is supported).

# How

Only change needed was to add tvOS to the podspec.

# Test Plan

- Manually test with a test app
- Add expo-linking to the modules included in the TV compile CI

# Checklist

- [x] Documentation is up to date to reflect these changes (eg: https://docs.expo.dev and README.md).
- [x] Conforms with the [Documentation Writing Style Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [x] This diff will work correctly for `npx expo prebuild` & EAS Build (eg: updated a module plugin).