[android] fix/update headless apploader to work on the new architecture (#32146)
# Why

On Android/newArch we're having issues with background tasks not working
correctly. Background tasks are used in `expo-background-fetch`,
`expo-location` and `expo-notifications` and are all driven by the
module `expo-task-manager` and its core
[TaskService](https://github.com/expo/expo/blob/031a3559dd669ed3edf1800a23eb1d86c52adb39/packages/expo-task-manager/android/src/main/java/expo/modules/taskManager/TaskService.java)
class on Android.

The `TaskService` class might be started from a background trigger like
a push notification being received by the device. In some cases the app
is not running, it might be stopped by the user, closed by the system or
just not being started after a device reboot. In such a case the service
will load the JS bundle in headless mode so that we can run code to make
network/api calls, update storage etc. from within the background. This
is handled by the
[RNHeadlessAppLoader](https://github.com/expo/expo/blob/37ec6ba3c191ae10b315789397b9428da2d48d03/packages/expo-modules-core/android/src/main/java/expo/modules/adapters/react/apploader/RNHeadlessAppLoader.kt)
class.

On Android/newArch we're seeing some exceptions when trying to run JS
code in headless mode:

```
Error: Exception in HostFunction: Could not enqueue microtask because they are disabled in this runtime, js engine: hermes
```

(We might also see other exceptions, but this is something I was
reliably seeing).

This seems to be caused by how we initialise our RN contexts when
running headless.

# How

I found some similar issues in the RN repo where they updated the
[HeadlessJSTaskService](https://github.com/robik/react-native/blob/ac870aa5e6a65f4253c2942aa1d8eaab408d1c90/packages/react-native/ReactAndroid/src/main/java/com/facebook/react/HeadlessJsTaskService.java):

[#45100](https://github.com/facebook/react-native/pull/45100)

This PR applies some of the same patterns for setting up the headless JS
system in Expo's `RNHeadlessAppLoader`.

# Test plan

Run background tests on Android using push-notifications, location and
background fetch on both new and old arch.

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>
Co-authored-by: Kudo Chien <kudo@expo.dev>