[android] added interface/helper for getting TaskService outside of a loaded app
In some instances `expo-task-manager` has tasks registered that are triggered by external events like remote notifications, background locations etc. In such an instance the app might not be running and will be started in the background without any activities visible.

The task manager has a feature where it can load an app in a headless js environment making it possible to run with everything needed to execute JS based tasks. For this to work the `TaskService` class needs to be instantiated since tasks are persisted to disk and are loaded by the `TaskService`.

The problem is that we don't want to add a hard dependency on `expo-task-manager` in our packages so we need some kind of lookup system to get the TaskService instance through its interface `TaskServiceInterface`.

Our current problem is that looking up instances by interfaces (`TaskManagerInterface` -> `TaskService`) through dependency injection is only possible when the module system has been loaded - which it is not in our use case.

The solution implemented here is to add a mechanism on the package level (which is available before the app is loaded) to retrieve the `TaskService` instance which is the class used for loading persisted tasks and finally load a headless environment to execute these tasks.

For libraries using `expo-task-manager` the way to use this feature is to make a simple call like the following:

```
TaskServiceProviderHelper.getTaskServiceImpl(context)
```

Where the context is the main application.

This is enough to instantiate and trigger the `TaskService` to load its persistent tasks and create the necessary consumers (in your code) and call them correctly.

It is not necesary to perform any checks as the `getTaskServiceImpl` method will use a static instance of the `TaskService` making the call non-intrusive. The same instance will be used by the module system when the app is loading so there should be no instance collisions by using this method.