[cli] add dedicated `/inspector/network` websocket endpoint (#32292)
# Why

Fusebox is great, although severely limited. There is no way for us to
hook into Fusebox on the native side -- and our PR to open this up just
a little bit has not been merged and will likely merge beyond React
Native 0.76.0 stable. Meaning, we have to work around this issue for
now.

# How

- Add `/inspector/network` websocket endpoint
- This websocket is used in the dev client to "emit" network related CDP
events
- These network CDP events are then re-broadcasted to _all connected
debuggers_.
- This is an issue when opening multiple debugger instances, but it's
the only way for us to hook into Fusebox for now.