[expo] allow dom component ref to use webview functions (#32419)
# Why

originally the `ref` to dom components is to the underlying webview.
after #31108, we does not pass ref to webview anymore. this pr tries to
fix the regression

# How

in the ref proxy object, we will pass to webview ref if the method is
not defined in `useDOMImperativeHandle`. to get the defined props from
`useDOMImperativeHandle`, we have to use postMessage to pass the defined
props from webview to native.