feat(expo router): add server action-only mode (#32432)
# Why

- To enable easier migration to server components we should add an
experimental mode which uses the classic router with parts of the new
router to enable server actions and React Navigation in parallel.
- The problem with this mode is that it only will have the same
server/client waterfalls as a standard React app. This should be fine in
the short-term.
- Users will also not be able to get full SSR/SSG with this mode as we
won't be able to server render routes.

<!--
Please describe the motivation for this PR, and link to relevant GitHub
issues, forums posts, or feature requests.
-->

# How

- Add `experiments.reactServerActions: true` which can be used with
`experiments.reactServerComponents: false` to render all routes in
client-mode while still allowing for importing server actions from
standalone files marked with "use server".
- The router imports and entry point will all automatically use the
classic router which means full native navigation. The output will still
need to have a server deployed to work in production.
- This new mode also fixes an issue with server actions not including
all possible client boundaries that are referenced in server actions.

<!--
How did you build this feature or fix this bug and why?
-->

# Test Plan

- Added two new e2e tests for dev and prod.
- Tested manually in router e2e.

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>