[fingerprint] add function signature stability test (#32493)
# Why

We want `eas-cli` to call the fingerprint library through the expo
package. This would be done by adding fingerprint to a dependency of the
expo package and reexporting it as expo/fingerprint like
[expo/config-plugins](https://github.com/expo/expo/blob/main/packages/expo/config-plugins.js).

In order to make sure that later versions of fingerprint dont get
released with an incompatible function signature, this PR adds a test in
the fingerprint package to make sure the function api doesnt change (or
is at least backwards compatible).

# Test Plan

- [ ] new tests pass

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
This is required for changes to Expo modules.
-->

- [ ] Documentation is up to date to reflect these changes (eg:
https://docs.expo.dev and README.md).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).