[autolinking] Handle package with no dependencies (#32547)
# Why

In testing UpdatesAPIDemo with SDK 52 beta, found that Expo autolinking crashed with `@expo-google-fonts/space-grotesk`. The resolver code hit a line where dependencies are expected to be defined, but this package has no dependencies.


# How

Added a check in the line that crashed, and issue is resolved.

# Test Plan

- Fix is tested in the SDK 52 branch of UpdatesAPIDemo
- We should add a unit test as well

# Checklist

<!--
Please check the appropriate items below if they apply to your diff. This is required for changes to Expo modules.
-->

- [x] Documentation is up to date to reflect these changes (eg: https://docs.expo.dev and README.md).
- [x] Conforms with the [Documentation Writing Style Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [x] This diff will work correctly for `npx expo prebuild` & EAS Build (eg: updated a module plugin).