[docs] Fix "The expected type comes from property 'ref'" error
While working through the tutorials I noticed an error in VSCode for `ref` in the `<View ref={imageRef}...` line. As someone new to Expo and React Native I was confused by this error. However, in a later code snippet the `imageRef` declaration changed and this fixed the error.

This commit makes the `imageRef` declaration consistent and removes the error.