[video] Add `audioMixingMode` property (#32428)
# Why

Right now we will always pause other audio in the system if the video is
outputting audio.

# How

I added a few options to allow the users to specify the desired
behaviour for the player.

- `mixWithOthers`: The player will mix its audio output with other apps.
- `duckOthers`: The player will lower the volume of other apps if any of
the active players is outputting audio.
- `auto`: The player will allow other apps to keep playing audio only
when it is muted.
- `doNotMix`: The player will pause playback in other apps, even when
it's muted.

where `auto` is the default and is basically the current behaviour. 
Also did a small fix in the NowPlayingManager. We have to call an update
every time the AudioSession changes to make sure it shows up correctly.
The whole thing is a bit messy and I'd like to rework it in the future
but we don't have the time for this now.

# Test Plan

Tested in BareExpo on Android and iOS