[iOS][gl] Convert GLView to Swift (#32675)
# Why

`EXGLView` is a legacy view, written in Objective-C. Our new
architecture interop for legacy views stopped working since #28829
As a result, `onSurfaceCreate` event handler was never set and so JS was
not receiving the GL object.

# How

Instead of fixing the interop for legacy views (expo-gl is the only
legacy view), I decided to convert it to Swift. I just converted the
code and renamed some properties. The implementation remains the same.

# Test Plan

Tested examples in bare-expo, on both the old and new architectures.
Note that some examples are not opening due to other unrelated issues
(in react-navigation and worklets integration).

Co-authored-by: Tomasz Sapeta <tsapeta@users.noreply.github.com>