[dev-launcher][ios] Fix handling deep links  (#32677)
# Why

Closes ENG-14002
Closes 32496

# How

With bridgeless mode on by default we can no longer rely on `[_appBridge
isValid]` to verify the state of the app, this PR updates the
dev-launcher logic to check for the `reactHost` instance in order to
determine if the app is running and then process the deep link. We we're
also not proving the correct `launchOptions` when recreating the
rootView

# Test Plan

FabricTester on iOS


https://github.com/user-attachments/assets/3efe3642-c2ae-46ed-8f86-573129cec199



# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
This is required for changes to Expo modules.
-->

- [ ] Documentation is up to date to reflect these changes (eg:
https://docs.expo.dev and README.md).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).